+++ app/Http/Controllers/Backend/RoleController.php
@@ -8,6 +8,7 @@ use App\Models\Module;
 use App\Models\Role;
 use Brian2694\Toastr\Facades\Toastr;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Str;
 
 class RoleController extends Controller
@@ -17,6 +18,8 @@ class RoleController extends Controller
      */
     public function index()
     {
+        Gate::authorize('index-role');
+
         $roles = Role::with(['permissions:id,permission_name,permission_slug'])
             ->select(['id', 'role_name', 'is_deleteable', 'updated_at'])
             ->get();
@@ -29,6 +32,7 @@ class RoleController extends Controller
      */
     public function create()
     {
+        Gate::authorize('create-role');
         $modules = Module::with(['permissions:id,module_id,permission_name,permission_slug'])->select(['id', 'module_name'])->get();
         return view('admin.pages.role.create', compact('modules'));
     }
@@ -38,6 +42,7 @@ class RoleController extends Controller
      */
     public function store(RoleStoreRequest $request)
     {
+        Gate::authorize('create-role');
         Role::updateOrCreate([
             'role_name' => $request->role_name,
             'role_slug' => Str::slug($request->role_name),
@@ -61,6 +66,7 @@ class RoleController extends Controller
      */
     public function edit(string $id)
     {
+        // Gate::authorize('edit-role');
         $role = Role::with(['permissions'])->find($id);
         $modules = Module::with(['permissions:id,module_id,permission_name,permission_slug'])->select(['id', 'module_name'])->get();
         return view('admin.pages.role.edit', compact('modules', 'role'));
@@ -71,6 +77,7 @@ class RoleController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        // Gate::authorize('edit-role');
         // $role = Role::with(['permissions'])->find($id);
         $role = Role::find($id);
         $role->update([
@@ -89,6 +96,7 @@ class RoleController extends Controller
      */
     public function destroy(string $id)
     {
+        Gate::authorize('delete-role');
         $role = Role::find($id);
         if ($role->is_deleteable) {
             $role->delete();
